requirements.txt
aiohttp==3.8.1
deta==1.1.0a2
Pillow==9.4.0
py-cord==2.3.2
matplotlib==3.6.2
numpy==1.23.5
pandas==1.5.2
